nixos/systemd-initrd: avoid system tainted message
systemd will print a "system tainted" warning message (`var-run-bad`) if
`/var/run` is not a symlink to `/run`, even in the initrd. To avoid
printing the warning message, create an empty `/run` and symlink
`/var/run` to it.

Also moves `/tmp` creation to `make-initrd-ng`, to avoid creating an
unnecessary store path for `/tmp/.keep`.